requirements.txt
# 项目依赖列表

# 核心框架
pytorch>=2.7.0
pytorch-cuda>=12.6
torchaudio>=2.7.0
torchvision>=0.18.0

# SAM3相关
sam3==0.1.0  # 注意：需要从官方获取访问权限

# 数据处理与图像操作
numpy>=2.0.0
pillow>=10.3.0
opencv-python>=4.10.0
six>=1.16.0

# 科学计算与数据可视化
scipy>=1.13.0
matplotlib>=3.9.0
seaborn>=0.13.2

# 进度显示
tqdm>=4.66.5

# 配置与日志
pyyaml>=6.0.1
logging>=0.4.9.6

# 数据集处理
dataset-tools>=0.1.0

# 可选：性能分析
psutil>=6.0.0
